pyproject.toml
[project]
name = "goit-pythonweb-hw-10"
version = "0.1.0"
description = ""
authors = [
    {name = "Artur Didur",email = "didur.art@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "greenlet (>=3.1.1,<4.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
]

[tool.poetry.scripts]
start = "src.app.main:main"

[tool.poetry]
packages = [{include = "app", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
